Add AppProvider and app context types to admin-x-framework
- Introduced AppProvider and related context types for managing app settings in Ghost Admin apps.
- Updated index.ts to export new AppProvider and context types.
- Refactored Posts and Stats apps to utilize the new AppProvider for app settings management.
