+++ apps/admin-x-framework/src/index.ts
@@ -2,6 +2,10 @@
 export type {StatsConfig, FrameworkContextType, FrameworkProviderProps, TopLevelFrameworkProps} from './providers/FrameworkProvider';
 export {FrameworkProvider, useFramework} from './providers/FrameworkProvider';
 
+// App Context
+export type {AppSettings, BaseAppProps, AppContextType, AppProviderProps} from './providers/AppProvider';
+export {AppProvider, useAppContext} from './providers/AppProvider';
+
 // Currency utilities
 export {getSymbol} from './utils/currency';
